add Open Source JWT to Apps, add validator to all controller, fix readme.

+++ app/Http/Controllers/VehicleController.php
@@ -27,29 +27,67 @@ class VehicleController extends Controller
                 'message' => 'no vehicles'
             ]);
         }
-        dd($vehicles);
     }
 
     public function show($id)
     {
+        $validator = \Validator::make(['id'=>$id], [
+            'id' => 'required|exists:vehicles,_id',
+        ],[
+            'id.exists' => 'id invalid'
+        ]);
+
+        if ($validator->fails()) {
+            return $validator->errors();
+        }
+
         $vehicle = $this->vehicleService->find($id);
         return response()->json($vehicle);
     }
 
     public function store(Request $request)
     {
+        $validator = \Validator::make($request->all(), [
+            'year' => 'required',
+            'color' => 'required' ,
+            'price' => 'required'
+        ]);
+
+        if ($validator->fails()) {
+            return $validator->errors();
+        }
         $vehicle = $this->vehicleService->create($request->all());
         return response()->json($vehicle, 201);
     }
 
     public function update(Request $request, $id)
     {
+        $validator = \Validator::make(['id'=>$id], [
+            'id' => 'required|exists:vehicles,_id',
+        ],[
+            'id.exists' => 'id invalid'
+        ]);
+
+        if ($validator->fails()) {
+            return $validator->errors();
+        }
+
         $vehicle = $this->vehicleService->update($id, $request->all());
         return response()->json($vehicle, 200);
     }
 
     public function destroy($id)
     {
+        $validator = \Validator::make(['id'=>$id], [
+            'id' => 'required|exists:vehicles,_id',
+        ],[
+            'id.exists' => 'id invalid'
+        ]);
+
+        if ($validator->fails()) {
+            return $validator->errors();
+        }
+
         $this->vehicleService->delete($id);
         return response()->json(null, 204);
     }
